+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Snyk Security Changelog
 
 ## [2.22.0]
-- Using Code Actions to view an AI Fix now shows the affected issue in a details pane.
+- Using Code Actions to view an AI Fix now shows the affected issue in the tree view and details pane.
 
 ## [2.21.0]
 - add scan summary